move `relengapi.db` to `relengapi.lib.db`
