+++ base/relengapi/blueprints/tokenauth/__init__.py
@@ -15,11 +15,11 @@ from flask.ext.login import login_required
 from itsdangerous import BadData
 from itsdangerous import JSONWebSignatureSerializer
 from relengapi import apimethod
-from relengapi import db
 from relengapi import p
 from relengapi.lib import angular
 from relengapi.lib import api
 from relengapi.lib import auth
+from relengapi.lib import db
 from relengapi.lib import permissions
 from werkzeug.exceptions import BadRequest
 from werkzeug.exceptions import NotFound
